Using frontend-build for prod build.
- Also adding “bugs” to package.json and updating the package-lock.
- The build.prod.js file has been renamed to create-apple-merchant-file and has been turned into a follow-up to the prod build, rather than a wrapper of it.
