+++ .env
@@ -1,0 +1,26 @@
+NODE_ENV='production'
+ACCESS_TOKEN_COOKIE_NAME=null
+BASE_URL=null
+CREDENTIALS_BASE_URL=null
+CSRF_TOKEN_API_PATH=null
+ECOMMERCE_BASE_URL=null
+LANGUAGE_PREFERENCE_COOKIE_NAME=null
+LMS_BASE_URL=null
+LOGIN_URL=null
+LOGOUT_URL=null
+MARKETING_SITE_BASE_URL=null
+ORDER_HISTORY_URL=null
+REFRESH_ACCESS_TOKEN_ENDPOINT=null
+SEGMENT_KEY=null
+SITE_NAME=null
+USER_INFO_COOKIE_NAME=null
+CURRENCY_COOKIE_NAME=null
+SUPPORT_URL=null
+CYBERSOURCE_URL=null
+APPLE_PAY_MERCHANT_NAME=null
+APPLE_PAY_COUNTRY_CODE=null
+APPLE_PAY_CURRENCY_CODE=null
+APPLE_PAY_START_SESSION_URL=null
+APPLE_PAY_AUTHORIZE_URL=null
+APPLE_PAY_SUPPORTED_NETWORKS=null
+APPLE_PAY_MERCHANT_CAPABILITIES=null
