+++ package.json
@@ -11,7 +11,7 @@
     "ie 11"
   ],
   "scripts": {
-    "build": "NODE_ENV=production BABEL_ENV=production node build.prod.js",
+    "build": "fedx-scripts webpack && node create-apple-merchant-file.js",
     "i18n_extract": "BABEL_ENV=i18n fedx-scripts babel src --quiet > /dev/null",
     "is-es5": "es-check es5 ./dist/*.js",
     "lint": "fedx-scripts eslint",
@@ -30,6 +30,9 @@
   "homepage": "https://github.com/edx/frontend-app-payment#readme",
   "publishConfig": {
     "access": "public"
+  },
+  "bugs": {
+    "url": "https://github.com/edx/frontend-app-payment/issues"
   },
   "dependencies": {
     "@edx/frontend-analytics": "3.0.0",
@@ -74,17 +77,26 @@
   "devDependencies": {
     "@edx/frontend-build": "1.1.0",
     "chance": "1.1.0",
+    "clean-webpack-plugin": "^3.0.0",
     "codecov": "3.6.1",
+    "dotenv-webpack": "^1.7.0",
     "enzyme": "3.10.0",
     "enzyme-adapter-react-16": "1.14.0",
     "es-check": "5.0.0",
     "glob": "7.1.4",
+    "html-webpack-plugin": "^3.2.0",
     "husky": "3.0.7",
     "jest": "24.9.0",
+    "lodash.defaultsdeep": "4.6.1",
+    "mini-css-extract-plugin": "^0.8.0",
+    "new-relic-source-map-webpack-plugin": "^1.2.0",
+    "purgecss-webpack-plugin": "1.6.0",
     "react-axe": "3.3.0",
     "react-test-renderer": "16.10.1",
     "reactifex": "1.1.1",
     "redux-mock-store": "1.5.3",
-    "rosie": "2.0.1"
+    "rosie": "2.0.1",
+    "webpack": "^4.41.0",
+    "webpack-bundle-analyzer": "^3.5.2"
   }
 }
